requirements.txt
Pillow==10.4.0
pandas==2.2.2
matplotlib==3.9.1
